+++ Package.swift
@@ -1,10 +1,11 @@
-// swift-tools-version: 5.6
+// swift-tools-version: 5.5
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
 
 let package = Package(
     name: "bytes",
+    defaultLocalization: "en",
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
@@ -22,8 +23,15 @@ let package = Package(
             name: "bytes",
             dependencies: [],
             path: "Sources"),
-        // .testTarget(
-        //     name: "bytesTests",
-        //     dependencies: ["bytes"]),
+        .testTarget(
+            name: "bytesTests",
+            dependencies: ["bytes"]),
     ]
 )
+
+#if swift(>=5.6)
+// Add the documentation compiler plugin if possible 
+package.dependencies.append(
+    .package(url: "https://github.com/apple/swift-docc-plugin", from: "1.0.0")
+)
+#endif